# backend/openapi.yaml
openapi: 3.1.0
info:
  title: Shion API
  version: 0.1.0
paths:
...
  /memories:
    get:
      summary: List memories of the logged-in user
      security:
        - cookieAuth: []
      responses:
        "200":
          description: List of memories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Memory"
        "401":
          description: Unauthorized
    post:
      summary: Create a new memory
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                note:
                  type: string
                location:
                  type: string
              required:
                - title
      responses:
        "201":
          description: Memory created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Memory"
        "401":
          description: Unauthorized
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: app_token
  schemas:
...
    Memory:
      type: object
      properties:
        id:
          type: integer
        created_by:
          type: integer
        title:
          type: string
        note:
          type: string
        location:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        deleted_at:
          type: string
          format: date-time
          nullable: true